requirements.txt
# Instagram API
instagrapi==2.1.2

# Database
sqlalchemy==2.0.23

# Telegram
python-telegram-bot==20.7

# Scheduling
schedule==1.2.0
APScheduler==3.10.4

# Utilities
python-dotenv==1.0.0
requests==2.31.0
pillow==10.1.0
pandas==2.1.3
python-dateutil==2.8.2
pytz==2023.3
cryptography==41.0.7
pysocks==1.7.1

# Monitoring
psutil==5.9.6

# Google Drive (opcional)
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0